added: CUDA parts

+++ basic/include/cuda.h
@@ -12,14 +12,14 @@ public:
 	}
 	virtual ~CUDA(){
 	}
-	void addition(double** sa1, double** sa2, double** da);
-	void multiplication(double** sa1, double** sa2, double** da);
-	void transpose(double** sa1, double** da);
+	void addition(int** sa1, int** sa2, int** da);
+	void multiplication(int** sa1, int** sa2, int** da);
+	void transpose(int** sa1, int** da);
 private:
 	int size;
 };
 
-void  CUDA::addition(double** sa1, double** sa2, double** da){
+void  CUDA::addition(int** sa1, int** sa2, int** da){
 	for(int i =0; i < size; i++){
 		for(int j =0; j < size; j++){
 			da[i][j] = sa1[i][j] + sa2[i][j];
@@ -27,7 +27,7 @@ void  CUDA::addition(double** sa1, double** sa2, double** da){
 	}
 }
 
-void CUDA::multiplication(double** sa1, double** sa2, double** da){
+void CUDA::multiplication(int** sa1, int** sa2, int** da){
 	for(int i = 0; i < size; i++){
 		for(int j = 0; j < size; j++){
 			da[i][j] = 0;			
@@ -38,7 +38,7 @@ void CUDA::multiplication(double** sa1, double** sa2, double** da){
 	}
 }
 
-void CUDA::transpose(double** sa1, double** da){
+void CUDA::transpose(int** sa1, int** da){
 	for(int i =0; i < size; i++){
 		for(int j =0; j < size; j++){
 			da[i][j] = sa1[j][i];
